+++ devel/0164.md
@@ -1,0 +1,53 @@
+# [0164] 迁移 (liii tree) 相关实现至 s7_liii_tree.c 和 s7_liii_tree.h
+
+## 任务相关的代码文件
+- src/s7.c
+- src/s7_internal.h
+- src/s7_internal_helpers.h
+- src/s7_liii_tree.h
+- src/s7_liii_tree.c
+- src/s7_scheme_predicate.h
+- src/s7_scheme_predicate.c
+- src/s7_liii_list.h
+- src/s7_liii_list.c
+- xmake.lua
+- tests/liii/tree/tree-memq-test.scm
+- devel/0164.md
+
+## 如何测试
+```bash
+xmake b goldfish
+bin/gf tests/liii/tree-test.scm
+bin/gf tests/liii/tree/tree-cyclic-p-test.scm
+bin/gf tests/liii/tree/tree-leaves-test.scm
+bin/gf tests/liii/tree/tree-memq-test.scm
+bin/gf tests/liii/tree/tree-set-memq-test.scm
+bin/gf tests/liii/tree/tree-count-test.scm
+bin/gf tests/scheme/s7-test.scm
+```
+
+## 2026-03-30 迁移 (liii tree) 相关实现到 s7_liii_tree.c
+
+### What
+1. 创建 `src/s7_liii_tree.h` 与 `src/s7_liii_tree.c`。
+2. 将 `s7.c` 中与 `(liii tree)` 相关的全部核心树函数迁移到 `src/s7_liii_tree.c`：
+   - `tree-cyclic?`: `tree_is_cyclic_or_has_pairs`、`tree_is_cyclic_1`、`tree_is_cyclic`
+   - `tree-leaves`: `tree_len_1`、`tree_len`、`tree_leaves_i_7p`
+   - `tree-memq`: `tree_memq_1`、`s7_tree_memq`、`tree_memq_b_7pp`、`tree_including_quote_memq`
+   - `tree-set-memq`: `pair_set_memq`、`tree_set_memq_b_7pp`、`tree_set_memq_syms_direct`、`tree_set_memq_chooser`
+   - `tree-count`: `tree_count`、`tree_count_at_least`、`g_tree_count`
+3. 统一收拢历史上暂存于其他模块的 tree 函数：
+   - 从 `src/s7_scheme_predicate.c` 迁移 `g_tree_is_cyclic`、`g_tree_leaves`、`g_tree_memq`、`g_tree_set_memq`、`g_tree_set_memq_syms`。
+   - 从 `src/s7_liii_list.c` 迁移 `tree_leaves_p_p`、`tree_set_memq_p_pp`。
+4. 在 `s7_internal_helpers.h` 中为 `small_symbol_set` 增加必要的桥接函数：`s7i_begin_small_symbol_set`、`s7i_end_small_symbol_set`、`s7i_add_symbol_to_small_symbol_set`、`s7i_symbol_is_in_small_symbol_set`。
+5. 在 `xmake.lua` 中配置 `src/s7_liii_tree.c`，并在 `src/s7.c` 与 `src/s7_internal.h` 中包含 `src/s7_liii_tree.h`。
+6. 完善 `tests/liii/tree/tree-memq-test.scm`，增加字符、布尔、整数、过程、复合对象（string/pair/vector）同一引用、点对 tail 等非 symbol 类型的全方位单测。
+
+### Why
+完成 s7.c 瘦身与模块化拆分，让 `(liii tree)` 的全部 C 实现（门面函数、优化器特化函数、核心树遍历算法）归拢于专用的 tree 模块中，消除跨模块历史杂质，降低各组件耦合。
+
+### How
+- `s7_liii_tree.c` 包含 `s7_internal.h`、`s7_liii_tree.h` 与 `s7_internal_helpers.h`。
+- `tree_is_cyclic`、`s7_tree_memq`、`tree_len`、`tree_count`、`tree_including_quote_memq` 由 `s7_liii_tree.h` 导出供 `s7.c` 调用。
+- `find_and_apply_method` 改为 extern 并在 `s7_internal.h` 声明，供方法重载分发使用。
+- `small_symbol_set` 的状态操作通过 `s7i_` 桥接封装。
