+++ src/s7_scheme_predicate.h
@@ -49,7 +49,7 @@ s7_pointer g_arity(s7_scheme *sc, s7_pointer args);
 s7_pointer g_is_c_pointer(s7_scheme *sc, s7_pointer args);
 s7_pointer g_is_openlet(s7_scheme *sc, s7_pointer args);
 s7_pointer g_is_funclet(s7_scheme *sc, s7_pointer args);
-s7_pointer g_tree_is_cyclic(s7_scheme *sc, s7_pointer args);
+/* g_tree_is_cyclic migrated to s7_liii_tree.h */
 s7_pointer g_type_of(s7_scheme *sc, s7_pointer args);
 s7_pointer g_is_eq(s7_scheme *sc, s7_pointer args);
 s7_pointer g_is_eqv(s7_scheme *sc, s7_pointer args);
@@ -63,15 +63,14 @@ s7_pointer g_c_object_type(s7_scheme *sc, s7_pointer args);
 s7_pointer g_c_object_let(s7_scheme *sc, s7_pointer args);
 s7_pointer g_c_pointer_weak1(s7_scheme *sc, s7_pointer args);
 s7_pointer g_c_pointer_weak2(s7_scheme *sc, s7_pointer args);
-s7_pointer g_tree_leaves(s7_scheme *sc, s7_pointer args);
+/* g_tree_leaves migrated to s7_liii_tree.h */
 s7_pointer g_cyclic_sequences(s7_scheme *sc, s7_pointer args);
 s7_pointer g_object_to_let(s7_scheme *sc, s7_pointer args);
 s7_pointer g_pair_line_number(s7_scheme *sc, s7_pointer args);
 s7_pointer g_port_line_number(s7_scheme *sc, s7_pointer args);
-s7_pointer g_tree_memq(s7_scheme *sc, s7_pointer args);
-s7_pointer g_tree_set_memq(s7_scheme *sc, s7_pointer args);
+/* g_tree_memq, g_tree_set_memq migrated to s7_liii_tree.h */
 s7_pointer g_format_nr(s7_scheme *sc, s7_pointer args);
-s7_pointer g_tree_set_memq_syms(s7_scheme *sc, s7_pointer args);
+/* g_tree_set_memq_syms migrated to s7_liii_tree.h */
 s7_pointer g_heap_analyze(s7_scheme *sc, s7_pointer args);
 s7_pointer g_show_op_stack(s7_scheme *sc, s7_pointer args);
 s7_pointer g_is_op_stack(s7_scheme *sc, s7_pointer args);
